==> 1.ETL/ETL_version1/00_Analisis_kpis.ipynb <==
import pandas as pd

# Diccionario de datos

## Dataset Energy

data = pd.read_csv('https://github.com/Wolfgang93/emissionsCO2/tree/main/1.ETL/rawDataset/energyco2.csv?raw=true')

data.head()

data.drop('Unnamed: 0', axis = 1, inplace = True)

data.info()

data = data[data['Year'] >= 1980]

data.info()

columns_dic = ['Country','Energy_type','Year','Energy_consumption','Energy_production','GDP','Population', 'Energy_intensity_per_capita', 'Energy_intensity_by_GDP', 'CO2_emission']

### ***Analisis de Country***

https://www.un.org/en/about-us/member-states#gotoT

countries = pd.read_clipboard()

***Consideramos solo miembros ONU*** :
 - Paises a eliminar: 'Antarctica', 'American Samoa', 'Bermuda', 'Burma', 'British Virgin Islands', 'Cayman Islands', 'Cook Islands', 'Falkland Islands', 'Faroe Islands', 'Guadeloupe', 'Guam', 'Greenland', 'Gibraltar', 'Hawaiian Trade Zone', 'Hong Kong', 'Kosovo', 'Macau', 'Martinique', 'Montserrat', 'Netherlands Antilles', 'New Caledonia', 'Niue', 'Northern Mariana Islands', 'Palestinian Territories', 'Puerto Rico', 'Reunion', 'Saint Helena', 'Saint Pierre and Miquelon', 'Taiwan', 'Turks and Caicos Islands', 'U.S. Pacific Islands', , 'U.S. Territories', 'U.S. Virgin Islands', 'Western Sahara', 'Wake Island', 

 - Paises faltantes: 'Andorra', 'Liechtenstein', 'Marshall Islands', 'Monaco', 'Myanmar', 'Palau', 'San Marino', 
 - Paises posibles a cambiar nombre: 'Brunei' ('Brunei Darussalam'), 'Gambia, The'(Gambia (Republic of The)), 'Moldova' (Republic of Moldova) , 

 - Paises a unir: ['Congo-Brazzaville', 'Congo-Kinshasa'], 
 - Generalidades: 'Former Yugoslavia' (Bosnia y Herzegovina, Croacia, Eslovenia, Macedonia, Montenegro y Serbia) , 'Former Serbia and Montenegro' (Serbia y Montenegro), 'Former Czechoslovakia' (Czech Republic and Slovakia), +Germany ('Germany, East','Germany, West'), 'World', 

data.Country[data.Country.str.match(r'(^[Mm].*)')==True].unique()

## Dataset Energy_Soure

data_2 = pd.read_csv('https://github.com/Wolfgang93/emissionsCO2/tree/main/1.ETL/rawDataset/owid-energy-consumption-source.csv?raw=true')

data_2.head()

data_2.country.unique().size

***Consideramos solo miembros ONU*** :
 - Paises a eliminar: 'Antarctica', 'American Samoa', 'British Virgin Islands', 'Cayman Islands', 'Chad', 'Cook Islands', 'Faeroe Islands', 'Falkland Islands', 'French Guiana', 'French Polynesia', 'G20 (Ember)', 'G7 (Ember)',  'Gibraltar', 'Greenland', 'Guadeloupe', 'Guam', 'Hong Kong', 'Kosovo', 'Macao', 'Martinique', 'Moldova', 'Montserrat', 'New Caledonia', 'Niue', 'Northern Mariana Islands', 'Puerto Rico', 'Palestine', 'Reunion', 'Saint Helena',  'Saint Pierre and Miquelon', 'Taiwan',  'Tanzania', 'Turks and Caicos Islands', 'Yugoslavia', 'Western Sahara'


 - Paises faltantes: 'Andorra',  'Belize', 'chad', 'Comoros', 'Liechtenstein', 'Marshall Islands', 'Monaco', 'Myanmar', 'Palau', 'San Marino', 

 - Paises posibles a cambiar nombre: 'Brunei' ('Brunei Darussalam'), 'Cape verde' (Cabo Verde), 'Guinea-Bissau' (Guinea Bissau), 'Timor' (Timor-Leste), 'Laos' (Lao People’s Democratic Republic), 'North Korea', 

 - Paises a unir: 

 - Generalidades: 'Africa', 'Africa (BP)', 'Africa (Shift)', 'Asia',
       'Asia & Oceania (EIA)', 'Asia Pacific (BP)',
       'Asia and Oceania (Shift)', 'Australia and New Zealand (EIA)', 'CIS (BP)', 'Central & South America (EIA)',
       'Central African Republic', 'Central America (BP)',
       'Central and South America (Shift)', 'Czechoslovakia', 'EU28 (Shift)', 'East Germany (EIA)', 'Eastern Africa (BP)', 'Eurasia (EIA)',
       'Eurasia (Shift)', 'Europe', 'Europe (BP)', 'Europe (Shift)',
       'European Union (27)', 'European Union (EIA)', 'Hawaiian Trade Zone (EIA)', 'High-income countries', 'IEO - Africa (EIA)', 'IEO - Middle East (EIA)',
       'IEO OECD - Europe (EIA)', 'Inde (Shift)', 'Latin America and Caribbean (Ember)', 'Low-income countries',
       'Lower-middle-income countries', 'Mexico, Chile, and other OECD Americas (EIA)',  'Middle Africa (BP)', 'Middle East (BP)',
       'Middle East (EIA)', 'Middle East (Ember)', 'Middle East (Shift)', 'Netherlands Antilles', 'Non-OECD (BP)', 'Non-OECD (EIA)', 'Non-OPEC (EIA)',
       'North America', 'North America (BP)', 'North America (Shift)' , 'OECD (BP)', 'OECD (EIA)', 'OECD (Ember)', 'OECD (Shift)',
       'OECD - Asia And Oceania (EIA)', 'OECD - Europe (EIA)',
       'OECD - North America (EIA)', 'OPEC (EIA)', 'OPEC (Shift)',
       'OPEC - Africa (EIA)', 'OPEC - South America (EIA)', 'Oceania', 'Other Africa (BP)', 'Other Asia Pacific (BP)',
       'Other CIS (BP)', 'Other Caribbean (BP)',
       'Other Eastern Africa (BP)', 'Other Europe (BP)',
       'Other Middle Africa (BP)', 'Other Middle East (BP)',
       'Other Non-OECD - America (EIA)', 'Other Non-OECD - Asia (EIA)',
       'Other Non-OECD - Europe and Eurasia (EIA)',
       'Other Northern Africa (BP)', 'Other South America (BP)',
       'Other South and Central America (BP)',
       'Other Southern Africa (BP)', 'Other Western Africa (BP)', 'Persian Gulf (EIA)', 'Persian Gulf (Shift)', 'Serbia and Montenegro', 'South America', 'South Korea and other OECD Asia (EIA)', 'South and Central America (BP)', 'Total Non-OPEC (BP)', 'Total OPEC (BP)', 'U.S. Pacific Islands (EIA)', 'U.S. Pacific Islands (Shift)',
       'U.S. Territories (EIA)', 'U.S. Territories (Shift)', 'USSR', 'United States Virgin Islands',
       'Upper-middle-income countries', 'Wake Island (EIA)', 'Wake Island (Shift)', 'West Germany (EIA)',
       'Western Africa (BP)', 'World'

data_2.country[data_2.country.str.match(r'(^[Zz].*)')==True].unique()


## Dataset Industries

data_3 = pd.read_csv('https://github.com/Wolfgang93/emissionsCO2/tree/main/1.ETL/rawDataset/global_power_plant_database.csv?raw=true')

data_3.head()

***Consideramos solo miembros ONU*** :
 - Paises a eliminar: 'Antarctica', 'American Samoa', 'French Guiana', 'Kosovo', 'Moldova', 'Palestine', 'Taiwan',  'Tanzania', 'Western Sahara'

 - Paises faltantes: 'Andorra',  'Belize', 'chad', 'Comoros', 'Liechtenstein', 'Marshall Islands', 'Monaco', 'Myanmar', 'Palau', 'San Marino', 

 - Paises posibles a cambiar nombre: 'Brunei' ('Brunei Darussalam'), 'Cape verde' (Cabo Verde), 'Laos' (Lao People’s Democratic Republic), 'North Korea', 'Swaziland' (Eswatini), 'Macedonia' (North Macedonia)

 - Paises a unir: 

 - Generalidades: 

data_3.country_long[data_3.country_long.str.match(r'(^[Tt].*)')==True].unique()

data_3.country_long.unique().size
