==> 1.ETL/ETL_version1/03_limpieza etl global_power_db.ipynb <==
## Proceso de limpieza dataset ***global_power***

Cargando csv global_power_plant_database

import pandas as pd

global_power_db = pd.read_csv('https://github.com/Wolfgang93/emissionsCO2/tree/main/1.ETL/rawDataset/global_power_plant_database.csv?raw=true', encoding= 'utf8')
global_power_db.head(5)

country / name energy plant / latitude longitude / primary fuel / other fuel 1/ other fuel 2/ other fuel 3/

##### ***Seleccionando las columnas que serán trabajadas***

global_power_db_filtred = global_power_db[["country",'country_long',"name","latitude","longitude","primary_fuel"]]
global_power_db_filtred.head(3)

Identificamos los campos nulos

global_power_db_filtred.info()

##### ***Revisando duplicados***

global_power_db_filtred.drop_duplicates(inplace= True)

global_power_db_filtred.info()

##### ***Normalización de columna países***

countries_list = ['Afghanistan', 'Albania', 'Algeria', 'American Samoa', 'Angola', 'Antarctica', 'Antigua and Barbuda', 'Argentina', 'Armenia', 'Aruba', 'Australia', 'Austria', 'Azerbaijan', 'Bahamas', 'Bahrain', 'Bangladesh', 'Barbados', 'Belarus', 'Belgium', 'Belize', 'Benin', 'Bermuda', 'Bhutan', 'Bolivia', 'Bosnia and Herzegovina', 'Botswana', 'Brazil', 'British Virgin Islands', 'Brunei', 'Bulgaria', 'Burkina Faso', 'Burundi', 'Cambodia', 'Cameroon', 'Canada', 'Cape Verde', 'Cayman Islands', 'Central African Republic', 'Chad', 'Chile', 'China', 'Colombia', 'Comoros', 'Congo', 'Cook Islands', 'Costa Rica', "Cote d'Ivoire", 'Croatia', 'Cuba', 'Cyprus', 'Czechia', 'Czechoslovakia', 'Democratic Republic of Congo', 'Denmark', 'Djibouti', 'Dominica', 'Dominican Republic', 'Ecuador', 'Egypt', 'El Salvador', 'Equatorial Guinea', 'Eritrea', 'Estonia', 'Eswatini', 'Ethiopia', 'Faeroe Islands', 'Falkland Islands', 'Fiji', 'Finland', 'France', 'French Guiana', 'French Polynesia', 'Gabon', 'Gambia', 'Georgia', 'Germany', 'Ghana', 'Gibraltar', 'Greece', 'Greenland', 'Grenada', 'Guadeloupe', 'Guam', 'Guatemala', 'Guinea', 'Guinea Bissau', 'Guyana', 'Haiti', 'Honduras', 'Hong Kong', 'Hungary', 'Iceland', 'India', 'Indonesia', 'Iran', 'Iraq', 'Ireland', 'Israel', 'Italy', 'Jamaica', 'Japan', 'Jordan', 'Kazakhstan', 'Kenya', 'Kiribati', 'Kosovo', 'Kuwait', 'Kyrgyzstan', 'Laos', 'Latvia', 'Lebanon', 'Lesotho', 'Liberia', 'Libya', 'Lithuania', 'Luxembourg', 'Macao', 'Madagascar', 'Malawi', 'Malaysia', 'Maldives', 'Mali', 'Malta', 'Martinique', 'Mauritania', 'Mauritius', 'Mexico', 'Micronesia', 'Moldova', 'Mongolia', 'Montenegro', 'Montserrat', 'Morocco', 'Mozambique', 'Myanmar', 'Namibia', 'Nauru', 'Nepal', 'Netherlands', 'Netherlands Antilles', 'New Caledonia', 'New Zealand', 'Nicaragua', 'Niger', 'Nigeria', 'Niue', 'North Korea', 'North Macedonia', 'Northern Mariana Islands', 'Norway', 'Oman', 'Pakistan', 'Palestine', 'Panama', 'Papua New Guinea', 'Paraguay', 'Peru', 'Philippines', 'Poland', 'Portugal', 'Puerto Rico', 'Qatar', 'Reunion', 'Romania', 'Russia', 'Rwanda', 'Saint Helena', 'Saint Kitts and Nevis', 'Saint Lucia', 'Saint Pierre and Miquelon', 'Saint Vincent and the Grenadines', 'Samoa', 'Sao Tome and Principe', 'Saudi Arabia', 'Senegal', 'Serbia', 'Seychelles', 'Sierra Leone', 'Singapore', 'Slovakia', 'Slovenia', 'Solomon Islands', 'Somalia', 'South Africa', 'South Korea', 'South Sudan', 'Spain', 'Sri Lanka', 'Sudan', 'Suriname', 'Sweden', 'Switzerland', 'Syria', 'Taiwan', 'Tajikistan', 'Tanzania', 'Thailand', 'Timor-Leste', 'Togo', 'Tonga', 'Trinidad and Tobago', 'Tunisia', 'Turkey', 'Turkmenistan', 'Turks and Caicos Islands', 'Tuvalu', 'Uganda', 'Ukraine', 'United Arab Emirates', 'United Kingdom', 'United States', 'United States Virgin Islands', 'Uruguay', 'Uzbekistan', 'Vanuatu', 'Venezuela', 'Vietnam', 'Western Sahara', 'Yemen',   'Yugoslavia', 'Zambia', 'Zimbabwe']

len(countries_list)

global_power_db_filtred = global_power_db_filtred.replace('Neuhardenberg PV3,4+5','Neuhardenberg PV3.4+5')

global_power_db_filtred = global_power_db_filtred.replace('Brunei Darussalam','Brunei')
global_power_db_filtred = global_power_db_filtred.replace('Cote DIvoire',"Cote d'Ivoire")
global_power_db_filtred = global_power_db_filtred.replace('Czech Republic','Czechia')
global_power_db_filtred = global_power_db_filtred.replace('Democratic Republic of the Congo','Congo')
global_power_db_filtred = global_power_db_filtred.replace('Macedonia', 'North Macedonia')
global_power_db_filtred = global_power_db_filtred.replace('Swaziland','Eswatini')
global_power_db_filtred = global_power_db_filtred.replace('Syrian Arab Republic', 'Syria')
global_power_db_filtred = global_power_db_filtred.replace('United States of America','United States')
global_power_db_filtred = global_power_db_filtred.replace('Guinea-Bissau','Guinea Bissau')


global_power_db_filtred_by_country = global_power_db_filtred[global_power_db_filtred.country_long.isin(countries_list)]

global_power_db_filtred

##### ***Normalización de columnas***

columns_global_plant = ['country','namePlant','latitude','longitude','primaryfuel']

global_power_db_filtred_by_country.drop('country', axis = 1, inplace=True)

global_power_db_filtred_by_country.columns = columns_global_plant

global_power_db_filtred_by_country.head()

##### Normalización de tabla ***global_power_db_filtred_by_country***


global_power_db_filtred_by_country.head()

global_power_db_filtred_by_country['primaryfuel'] = global_power_db_filtred_by_country['primaryfuel'].str.lower()

global_power_db_filtred_by_country['primaryfuel'].unique()

***Tabla plant***

plant = global_power_db_filtred_by_country[['country','namePlant','latitude','longitude','primaryfuel']]

plant.drop_duplicates(subset=['country','namePlant'], inplace=True)

plant.head()

plant.sort_values(by='country').reset_index(drop=True).to_csv('../Modelo_ER/plant.csv', header=True, index=False)

##### ***Subida a github***

#from github import Github

#token = ''

#g = Github(token)

#repo = g.get_repo('')

#Crear/Actualizar csv  en el repo git
#try:
#    repo.create_file('global_power_db_filtred_by_country.csv','Limpieza plantas por país', global_power_db_filtred.to_csv(index = False))
#except:
#    contents = repo.get_contents('global_power_db_filtred_by_country.csv')
#    repo.update_file(contents.path, 'Actualización, limpieza plantas por país', global_power_db_filtred.to_csv(index = False), contents.sha)

global_power_db_filtred.to_csv('../normalizedDataset/global_power_db_filtred_by_country.csv',index = False)
