==> 1.ETL/ETL_version1/04_analisis_groups.ipynb <==
# Groups

import pandas as pd
import numpy as np

#### Tabla ***region***

groups = ['Africa','Asia','Central America','Europe','North America','Oceania','South America', 'World','G20 (Ember)']

region = pd.DataFrame(groups, columns=['region']).sort_values(by='region').reset_index().rename(columns={'index':'idRegion'})

region.to_csv('../Modelo_ER/region.csv', index=False)

#### Tabla ***emission_region***


energy_consumption = pd.read_csv("https://github.com/Wolfgang93/emissionsCO2/tree/main/1.ETL/rawDataset/owid-energy-consumption-source.csv?raw=true")
owid_co2_data = pd.read_csv('../rawDataset/owid-co2-data.csv')
countries = pd.read_csv('../Modelo_ER/country.csv').country.values
regions = pd.read_csv('../Modelo_ER/region.csv').region.values

##### Procesamiento energy_consumption

quitar_columnas = ['biofuel_cons_change_pct','biofuel_cons_change_twh','biofuel_cons_per_capita','biofuel_consumption','biofuel_elec_per_capita','biofuel_electricity','biofuel_share_elec','biofuel_share_energy','coal_cons_change_pct','coal_cons_change_twh','coal_cons_per_capita','coal_prod_change_pct','coal_prod_change_twh','coal_prod_per_capita','coal_production','coal_share_elec','coal_share_energy','energy_cons_change_pct','energy_cons_change_twh','fossil_cons_change_pct','fossil_cons_change_twh','fossil_elec_per_capita','fossil_electricity','fossil_energy_per_capita','fossil_fuel_consumption','fossil_share_elec','fossil_share_energy','gas_cons_change_pct','gas_cons_change_twh','gas_prod_change_pct','gas_prod_change_twh','gas_prod_per_capita','gas_production','gas_share_elec','gas_share_energy','hydro_cons_change_pct','hydro_cons_change_twh','hydro_share_elec','hydro_share_energy','low_carbon_cons_change_pct','low_carbon_cons_change_twh','low_carbon_consumption','low_carbon_elec_per_capita','low_carbon_electricity','low_carbon_energy_per_capita','low_carbon_share_elec','low_carbon_share_energy','net_elec_imports','net_elec_imports_share_demand','nuclear_cons_change_pct','nuclear_cons_change_twh','nuclear_share_elec','nuclear_share_energy','oil_cons_change_pct','oil_cons_change_twh','oil_prod_change_pct','oil_prod_change_twh','oil_prod_per_capita','oil_production','oil_share_elec','oil_share_energy','other_renewable_exc_biofuel_electricity','other_renewables_cons_change_pct','other_renewables_cons_change_twh','other_renewables_elec_per_capita_exc_biofuel','other_renewables_share_elec','other_renewables_share_elec_exc_biofuel','other_renewables_share_energy','renewables_cons_change_pct', 'renewables_cons_change_twh','renewables_share_elec','renewables_share_energy','solar_cons_change_pct','solar_cons_change_twh','solar_share_elec','solar_share_energy','wind_cons_change_pct','wind_cons_change_twh','wind_share_elec','wind_share_energy','energy_per_gdp', 'gdp', 'primary_energy_consumption']

energy_consumption = energy_consumption.drop(quitar_columnas, axis=1)
energy_consumption.drop_duplicates(inplace=True)
energy_consumption = energy_consumption.fillna(0)
energy_consumption = energy_consumption[energy_consumption['year'] >= 1980]
energy_consumption = energy_consumption.replace(',', ' ', regex=True)
energy_consumption['population'] = energy_consumption['population'].astype(int)
energy_consumption = energy_consumption.replace('Central America (BP)','Central America')

energy_consumption_region = energy_consumption[energy_consumption.country.isin(regions)]

energy_consumption_region = energy_consumption_region[['country', 'year', 'population', 'greenhouse_gas_emissions']]

energy_consumption_region = energy_consumption_region.rename(columns={'greenhouse_gas_emissions':'co2emissions', 'country':'region'})

##### Procesamiento owid_co2_data

cols = ['country', 'year', 'co2_per_capita']
owid_co2_data = owid_co2_data[cols]
owid_co2_data = owid_co2_data[owid_co2_data['year'] >= 1980]
owid_co2_data.columns = ['region', 'year', 'co2PerCapita']
owid_co2_data = owid_co2_data.replace('Micronesia (country)','Micronesia')
owid_co2_data = owid_co2_data.replace('Timor','Timor-Leste')
owid_co2_data = owid_co2_data.replace('Guinea-Bissau','Guinea Bissau')

owid_co2_regions = owid_co2_data[owid_co2_data['region'].isin(regions)]

##### Merge de ambos DF finales de los procesamientos

emission_region = energy_consumption_region.merge(owid_co2_regions, on=['region', 'year'], how = 'left').fillna(0).sort_values(['region', 'year'])

emission_region

emission_region.to_csv('../Modelo_ER/emission_region.csv', index=False)

#### Tabla ***source_region_electricity***


##### Procesamiento

energy_consumption = energy_consumption.rename(columns={'other_renewable_consumption':'other_renewables_consumption','other_renewable_electricity':'other_renewables_electricity'})

energy_consumption['coal_energy_per_capita'] = np.zeros(len(energy_consumption))

energy_consumption = energy_consumption.rename(columns={'other_renewable_consumption':'other_renewables_consumption','other_renewable_electricity':'other_renewables_electricity'})

source_region_electricity = energy_consumption.drop(['iso_code','population','carbon_intensity_elec','electricity_demand','electricity_generation','energy_per_capita','greenhouse_gas_emissions'], axis=1)

source_region_electricity = source_region_electricity[source_region_electricity.country.isin(regions)]

source_region_electricity

##### Pivot de las columnas

sources = ['coal', 'gas', 'hydro', 'nuclear', 'oil', 'other_renewables', 'renewables', 'solar', 'wind']
cols = ['region','year', 'source', 'consumptionElectricity', 'productionElectricity', 'energyPerCapita', 'elecPerCapita']

base_df = pd.DataFrame(columns = cols)

for n in sources:
    agg = source_region_electricity[['country', 'year', f"{n}_consumption", f"{n}_electricity", f"{n}_energy_per_capita", f"{n}_elec_per_capita"]]
    agg.columns = ['region','year','consumptionElectricity', 'productionElectricity', 'energyPerCapita', 'elecPerCapita']
    agg.insert(2, 'source', n)
    base_df = base_df.append(agg)

base_df.drop_duplicates(subset=['region', 'year', 'source'], inplace=True)
base_df.sort_values(by=['region','year','source'], inplace=True)

base_df

base_df.to_csv('../Modelo_ER/source_region_electricity.csv', index=False)
